Refactors ViewRequestModal to separate view request body from modal, to make cart easier in the future
